feat: Add unit tests to confirm identifier escaping functionality
I also needed to rename the file so I'm not misspelling the word manifest.
